circle ci env

+++ src/snovault/tests/serverfixtures.py
@@ -1,5 +1,5 @@
 import pytest
-
+import os
 
 def pytest_configure():
     import logging
@@ -24,35 +24,41 @@ def pytest_configure():
 @pytest.yield_fixture(scope='session')
 def engine_url(request):
     # Ideally this would use a different database on the same postgres server
-    from urllib.parse import quote
-    from .postgresql_fixture import initdb, server_process
-    tmpdir = request.config._tmpdirhandler.mktemp('postgresql-engine', numbered=True)
-    tmpdir = str(tmpdir)
-    initdb(tmpdir)
-    process = server_process(tmpdir)
+    if os.environ.get("CIRCLECI") or True:
+        yield 'postgresql://postgres@localhost:5432/postgres'
+    else:
+        from urllib.parse import quote
+        from .postgresql_fixture import initdb, server_process
+        tmpdir = request.config._tmpdirhandler.mktemp('postgresql-engine', numbered=True)
+        tmpdir = str(tmpdir)
+        initdb(tmpdir)
+        process = server_process(tmpdir)
 
-    yield 'postgresql://postgres@:5432/postgres?host=%s' % quote(tmpdir)
+        yield 'postgresql://postgres@:5432/postgres?host=%s' % quote(tmpdir)
 
-    if process.poll() is None:
-        process.terminate()
-        process.wait()
+        if process.poll() is None:
+            process.terminate()
+            process.wait()
 
 
 @pytest.mark.fixture_cost(10)
 @pytest.yield_fixture(scope='session')
 def postgresql_server(request):
-    from urllib.parse import quote
-    from .postgresql_fixture import initdb, server_process
-    tmpdir = request.config._tmpdirhandler.mktemp('postgresql', numbered=True)
-    tmpdir = str(tmpdir)
-    initdb(tmpdir)
-    process = server_process(tmpdir)
+    if os.environ.get("CIRCLECI") or True:
+        yield 'postgresql://postgres@localhost:5432/postgres'
+    else:
+        from urllib.parse import quote
+        from .postgresql_fixture import initdb, server_process
+        tmpdir = request.config._tmpdirhandler.mktemp('postgresql', numbered=True)
+        tmpdir = str(tmpdir)
+        initdb(tmpdir)
+        process = server_process(tmpdir)
 
-    yield 'postgresql://postgres@:5432/postgres?host=%s' % quote(tmpdir)
+        yield 'postgresql://postgres@:5432/postgres?host=%s' % quote(tmpdir)
 
-    if process.poll() is None:
-        process.terminate()
-        process.wait()
+        if process.poll() is None:
+            process.terminate()
+            process.wait()
 
 
 @pytest.fixture(scope='session')
@@ -64,17 +70,20 @@ def elasticsearch_host_port():
 @pytest.mark.fixture_cost(10)
 @pytest.yield_fixture(scope='session')
 def elasticsearch_server(request, elasticsearch_host_port):
-    from .elasticsearch_fixture import server_process
-    host, port = elasticsearch_host_port
-    tmpdir = request.config._tmpdirhandler.mktemp('elasticsearch', numbered=True)
-    tmpdir = str(tmpdir)
-    process = server_process(str(tmpdir), host=host, port=port)
+    if os.environ.get("CIRCLECI") or True:
+        yield 'http://localhost:9200'
+    else:
+        from .elasticsearch_fixture import server_process
+        host, port = elasticsearch_host_port
+        tmpdir = request.config._tmpdirhandler.mktemp('elasticsearch', numbered=True)
+        tmpdir = str(tmpdir)
+        process = server_process(str(tmpdir), host=host, port=port)
 
-    yield 'http://%s:%d' % (host, port)
+        yield 'http://%s:%d' % (host, port)
 
-    if process.poll() is None:
-        process.terminate()
-        process.wait()
+        if process.poll() is None:
+            process.terminate()
+            process.wait()
 
 
 # http://docs.sqlalchemy.org/en/rel_0_8/orm/session.html#joining-a-session-into-an-external-transaction
